+++ src/main/java/com/barlinc/unusual_prehistory/datagen/UP2BlockstateProvider.java
@@ -51,7 +51,7 @@ public class UP2BlockstateProvider extends BlockStateProvider {
         this.tallPlant(LARGE_HORSETAIL);
         this.tallPlant(RAIGUENRAYUN);
         this.tallPlant(ZHANGSOLVA_BLOOM);
-        this.tallPlant(DELITZSCHALA_FERN);
+        this.tallPlant(DELITZSCHALA_STALK);
 
         // dryophyllum
         this.pillar(DRYOPHYLLUM_LOG);